+++ lib/src/lexemes/optional.dart
@@ -6,7 +6,7 @@ class OptionalLexeme extends PatternLexeme {
   OptionalLexeme(super.pattern, { super.name, super.grammar });
 
   @override
-  Token tokenize(String string, [int start = 0]) {
+  Token tokenize(String string, [ int start = 0 ]) {
     final token = pattern.optionalTokenizeFrom(this, string, start);
     if (token == null) return Token.emptyAt(this, string, start);
     return Token.match(this, token);